// packages/components/orchestration/protocol/src/agents/types.ts
import type { JubjubPoint } from "@midnight-ntwrk/compact-runtime";
import type { VerificationMethodRef } from "@midnight-ntwrk/midnight-did-credentials/managed/credentials/contract/index.js";

export type Signer = {
  readonly label: string;
  readonly secretKey: bigint;
  readonly publicKey: JubjubPoint;
  readonly verificationMethodRef: VerificationMethodRef;
};

export type PartyRole = "issuer" | "holder" | "verifier";

export type DIDProfile = {
  readonly role: PartyRole;
  readonly label: string;
  readonly signer: Signer;
};